Moved ResourceLoader to common libs
