+++ apps/hops-api/src/test/kotlin/no/nav/helse/hops/testUtils/TestContainerFactory.kt
@@ -2,17 +2,18 @@ package no.nav.helse.hops.testUtils
 
 import org.testcontainers.containers.GenericContainer
 import org.testcontainers.containers.wait.strategy.Wait
+import java.time.Duration
 
 object TestContainerFactory {
     fun hapiFhirServer() = GenericContainer<Nothing>(
         "hapiproject/hapi:v5.3.0"
     ).apply {
         withExposedPorts(8080)
-        waitingFor(Wait.forHttp("/fhir/metadata"))
+        waitingFor(Wait.forHttp("/fhir/metadata").withStartupTimeout(Duration.ofMinutes(2)))
     }
 
     fun mockOauth2Server() = GenericContainer<Nothing>(
-        "docker.pkg.github.com/navikt/mock-oauth2-server/mock-oauth2-server:0.3.2"
+        "ghcr.io/navikt/mock-oauth2-server:0.3.3"
     ).apply {
         withEnv("SERVER_PORT", "8081")
         withExposedPorts(8081)
