+++ apps/hops-bestilling/src/test/kotlin/no/nav/helse/hops/infrastructure/FhirResourceValidatorHapiTest.kt
@@ -2,8 +2,8 @@ package no.nav.helse.hops.infrastructure
 
 import kotlinx.coroutines.runBlocking
 import no.nav.helse.hops.domain.isAllOk
+import no.nav.helse.hops.fhir.FhirResourceLoader
 import no.nav.helse.hops.fhir.toJson
-import no.nav.helse.hops.testUtils.ResourceLoader
 import org.hl7.fhir.r4.model.Bundle
 import org.junit.jupiter.api.Test
 
@@ -12,7 +12,7 @@ internal class FhirResourceValidatorHapiTest {
     fun `validate valid message bundle`() {
         val sut = FhirResourceValidatorHapi
 
-        val message = ResourceLoader.asFhirResource<Bundle>("/fhir/valid-message.json")
+        val message = FhirResourceLoader.asResource<Bundle>("/fhir/valid-message.json")
         val result = runBlocking { sut.validate(message) }
 
         kotlin.test.assertTrue(result.isAllOk(), result.toJson())
@@ -22,7 +22,7 @@ internal class FhirResourceValidatorHapiTest {
     fun `validate message bundle without required message-header`() {
         val sut = FhirResourceValidatorHapi
 
-        val message = ResourceLoader.asFhirResource<Bundle>("/fhir/invalid-message-missing-header.json")
+        val message = FhirResourceLoader.asResource<Bundle>("/fhir/invalid-message-missing-header.json")
         val result = runBlocking { sut.validate(message) }
 
         kotlin.test.assertFalse(result.isAllOk(), result.toJson())
