--- apps/hops-hendelser/src/test/kotlin/no/nav/helse/hops/testUtils/ResourceLoader.kt
@@ -1,23 +1,0 @@
-package no.nav.helse.hops.testUtils
-
-import no.nav.helse.hops.fhir.JsonConverter
-import org.hl7.fhir.instance.model.api.IBaseResource
-
-internal object ResourceLoader {
-    fun asString(resource: String): String =
-        try {
-            object {}.javaClass.getResource(resource)!!.readText(Charsets.UTF_8)
-        } catch (all: Exception) {
-            throw RuntimeException("Failed to load resource=$resource!", all)
-        }
-
-    fun asByteArray(resource: String): ByteArray =
-        try {
-            object {}.javaClass.getResource(resource)!!.readBytes()
-        } catch (all: Exception) {
-            throw RuntimeException("Failed to load resource=$resource!", all)
-        }
-
-    inline fun <reified R : IBaseResource> asFhirResource(resource: String) =
-        JsonConverter.parse<R>(asString(resource))
-}
