+++ libs/hops-common-core/src/main/kotlin/no/nav/helse/hops/ResourceLoader.kt
@@ -1,9 +1,7 @@
-package no.nav.helse.hops.testUtils
+package no.nav.helse.hops
 
-import no.nav.helse.hops.fhir.JsonConverter
-import org.hl7.fhir.instance.model.api.IBaseResource
-
-internal object ResourceLoader {
+object ResourceLoader {
+    /** Loads a file from the class-path as a String, this includes files in the /resources dir. **/
     fun asString(resource: String): String =
         try {
             object {}.javaClass.getResource(resource)!!.readText(Charsets.UTF_8)
@@ -11,13 +9,11 @@ internal object ResourceLoader {
             throw RuntimeException("Failed to load resource=$resource!", all)
         }
 
+    /** Loads a file from the class-path as a ByteArray, this includes files in the /resources dir. **/
     fun asByteArray(resource: String): ByteArray =
         try {
             object {}.javaClass.getResource(resource)!!.readBytes()
         } catch (all: Exception) {
             throw RuntimeException("Failed to load resource=$resource!", all)
         }
-
-    inline fun <reified R : IBaseResource> asFhirResource(resource: String) =
-        JsonConverter.parse<R>(asString(resource))
 }
